Change long types to int.
On the neo-vm int or long are both handled as BigInteger, i.e., it makes sense to have one standard number type in Java smart contracts, which is `int`..

+++ devpack/src/main/java/io/neow3j/devpack/neo/Blockchain.java
@@ -15,7 +15,7 @@ import io.neow3j.devpack.annotations.Syscall;
 public class Blockchain {
 
     @Syscall(SYSTEM_BLOCKCHAIN_GETHEIGHT)
-    public static native long getHeight();
+    public static native int getHeight();
 
     /**
      * Gets the block at the given block height.
@@ -24,7 +24,7 @@ public class Blockchain {
      * @return the <tt>Block</tt>.
      */
     @Syscall(SYSTEM_BLOCKCHAIN_GETBLOCK)
-    public static native Block getBlock(long height);
+    public static native Block getBlock(int height);
 
     /**
      * Gets the block with the given block hash.
@@ -62,7 +62,7 @@ public class Blockchain {
      * @return the <tt>Transaction</tt>.
      */
     @Syscall(SYSTEM_BLOCKCHAIN_GETTRANSACTIONFROMBLOCK)
-    public static native Transaction getTransactionFromBlock(long blockHeight, int txIndex);
+    public static native Transaction getTransactionFromBlock(int blockHeight, int txIndex);
 
     /**
      * Gets the transaction height of the transaction with the given hash. The transaction
